move flash function to its own file and include it to dependencies
