+++ helpers/html.php
@@ -233,14 +233,4 @@
         }
         return $array; 
     }
-
-    function flash( $message, $type = 'error' ) {
-        switch ( $type ) {
-            case 'error':
-            case 'warning':
-            case 'success':
-                $_SESSION[ 'alert' ] = [ $message, $type ];
-                break;
-        }
-    }
 ?>
